Replaced the complex env for a simpler one with a enum.  The complex one is getting removed as the API will not add things to the db on startup anymore.
